Remove old parse_game/enhance
