+++ gd2score/parse_game.py
@@ -1,87 +1,113 @@
-import re
-import xml.etree.ElementTree as ET
+import statsapi
 
-from .models import Game, Inning, HalfInning, Action, AtBat, Runner
-
-BASES = ('1B', '2B', '3B')
+from .models import Game, Inning, HalfInning, Action, AtBat, Runner, Player
+from .enhance import GameEnhancer
+from .draw_scorecard import DrawScorecard
 
 
 class GameParser:
-    def parse(self, xml):
+    def parse(self, game_dict):
+        game_id = game_dict['game_id']
+
         game = Game()
+        # Used later for logo lookups
+        game.away = self.get_team_code(game_dict['away_id'])
+        game.home = self.get_team_code(game_dict['home_id'])
 
-        game_root = ET.fromstring(xml)
-        ind = game_root.attrib['ind']
-        if ind == 'I':
-            game.in_progress = True
+        active_inning = None
+        active_half = None
+        top_bottom = 'top'
 
-        for inning in game_root:
-            game.add_inning(self.parse_inning(inning))
+        pbp = statsapi.get('game_playByPlay', {'gamePk': game_id})
+        for play in pbp['allPlays']:
+            inning_num = int(play['about']['inning'])
+            if not active_inning or active_inning.num != inning_num:
+                active_inning = Inning(inning_num)
+                game.add_inning(active_inning)
+
+            if not active_half or top_bottom != play['about']['halfInning']:
+                top_bottom = play['about']['halfInning']
+                active_half = HalfInning()
+                active_inning.add_half(active_half)
+
+            ab = AtBat(int(play['atBatIndex'] + 1),
+                       int(play['atBatIndex'] + 1),
+                       int(play['matchup']['batter']['id']),
+                       play['result']['description'],
+                       play['result']['event'],
+                       int(play['matchup']['pitcher']['id']),
+                       int(play['count']['outs']),
+                       int(play['result']['homeScore']),
+                       int(play['result']['awayScore']))
+
+            self.parse_runners(ab, play)
+            active_half.add_atbat(ab)
+
         return game
 
-    def parse_inning(self, inning):
-        active_inning = Inning(float(inning.attrib['num']))
-        for half_inning in inning:
-            active_inning.add_half(self.parse_half_inning(half_inning))
-        return active_inning
+    def parse_runners(self, ab, play):
+        runner_index = set(play['runnerIndex'])
+        for i, runner in enumerate(play['runners']):
+            mvmt = runner['movement']
+            start = self.get_base(mvmt['start'])
+            end = self.get_base(mvmt['end'])
+            if not start and not end:
+                continue
+            if mvmt['outBase']:
+                end = self.get_base(mvmt['outBase'])
 
-    def parse_half_inning(self, half_inning):
-        half = HalfInning()
-        for event in half_inning:
-            if event.tag == 'atbat':
-                half.add_atbat(self.parse_atbat(event))
-            elif event.tag == 'action':
-                half.add_action(self.parse_action(event))
-            else:
-                raise Exception('Unknown event type')
-        return half
+            # Runners that advance multiple bases are listed for each base they
+            # advance. Only create one Runner object per person.
+            already_added = False
+            play_index = int(runner['details']['playIndex'])
+            runner_id = runner['details']['runner']['id']
+            for ab_runner in ab.runners:
+                if (ab_runner.id == runner_id and
+                    ab_runner.event_num == play_index):
+                        assert end > ab_runner.end
+                        ab_runner.end = end  # Increase end
+                        already_added = True
+                        break
 
-    def parse_atbat(self, atbat):
-        pa_num = int(atbat.attrib['num'])
-        event_num = int(atbat.attrib['event_num'])
-        batter = int(atbat.attrib['batter'])
-        des = re.sub('\s\s+', ' ', atbat.attrib['des'].strip())
-        event = atbat.attrib['event']
-        outs = int(atbat.attrib['o'])
-        home_score = int(atbat.attrib['home_team_runs'])
-        away_score = int(atbat.attrib['away_team_runs'])
-        pitcher = int(atbat.attrib['pitcher'])
+            if not already_added:
+                r = Runner(runner_id, start, end, play_index)
+                r.to_score = runner['details']['isScoringEvent']
+                r.out = mvmt['isOut']
 
-        active_atbat = AtBat(pa_num, event_num, batter, des, event,
-                             pitcher, outs,
-                             home_score=home_score, away_score=away_score)
+                if i in runner_index:
+                    ab.add_atbat_runner(r)
+                else:
+                    ab.add_mid_pa_runner(r)
 
-        for child in atbat:
-            if child.tag == 'runner':
-                active_atbat.add_runner(self.parse_runner(child))
-            elif child.tag in ['pitch', 'po']:
-                pass
-            else:
-                raise Exception('Unknown atbat type: %s' % child.tag)
-
-        return active_atbat
-
-    def parse_action(self, action):
-        event_num = int(action.attrib['event_num'])
-        event = action.attrib['event']
-        des = re.sub('\s+', ' ', action.attrib['des']).strip()
-        player_id = int(action.attrib['player'])
-        return Action(event_num, event, des, player_id)
-
-    def parse_runner(self, runner):
-        id = int(runner.attrib['id'])
-        start = self.parse_base(runner.attrib['start'])
-        end = self.parse_base(runner.attrib['end'])
-        event_num = int(runner.attrib['event_num'])
-
-        if runner.attrib.get('score') == 'T':  # Scores!
-            end = 4
-        return Runner(id, start, end, event_num)
-
-    def parse_base(self, base):
+    def get_base(self, base):
         if not base:
             return 0
-        elif base in BASES:
-            return int(base[0])
+        elif base == 'score':
+            return 4
         else:
-            raise Exception('Could not parse base: %s' % base)
+            return int(base[0])
+
+    def get_team_code(self, team_id):
+        team = statsapi.get('team', {'teamId': team_id})['teams'][0]
+        return team['teamCode']
+
+
+if __name__ == '__main__':
+    games = statsapi.schedule(start_date='10/26/2018')
+    game_dict = games[0]
+
+    game = GameParser().parse(game_dict)
+    GameEnhancer().execute(game)
+    drawing = DrawScorecard().draw(game)
+    drawing.saveas('test.svg')
+
+    for inning in game:
+        print(inning)
+        for half in inning:
+            print(half)
+            for atbat in half:
+                print(atbat)
+                for runner in atbat.mid_pa_runners:
+                    print(runner, 'mid-pa')
+                for runner in atbat.runners:
+                    print(runner)
